debugger/requirements.txt
pylint==3.0.3
astroid==3.0.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
pytest==8.0.0
black==24.1.1
isort==5.13.2
colorama==0.4.6
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.1.0 